refactor: move out ImagePreviewModal to reuse
